Use ApplyBlockExpression to wrap all apply contents
While checking the AST, we want to be able to identify the contents of
the apply block and handle it separately. To do that we wrap it in an
ApplyBlockExpression. Later when passing it on to the `apply_executor`,
create a block or expression and hide the details of
ApplyBlockExpression.